Refactor runtime callbacks into .cc files
